commit-checker: move to an importable dir
