+++ commitchecker/commitchecker/git.go
@@ -1,4 +1,4 @@
-package main
+package commitchecker
 
 import (
 	"bytes"
